requirements.txt
alembic==1.13.3
annotated-types==0.7.0
bcrypt==4.2.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
mysqlclient==2.2.4
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
SQLAlchemy==2.0.36
uvicorn==0.32.0
sqlalchemy-utils==0.41.2